chore: refactored database into sub-components
